requirements.txt
numpy
pandas
scipy
matplotlib
biopython
numba
joblib
pyyaml
cupy-cuda101  # Adjust based on your CUDA version. Remove if not using NVIDIA GPUs.
